DailyFresh/requirements.txt
amqp==2.1.4
appdirs==1.4.3
billiard==3.5.0.2
celery==4.0.2
certifi==2017.7.27.1
chardet==3.0.4
Django==1.8.7
django-haystack==2.6.0
django-redis==4.8.0
django-tinymce==2.6.0
fdfs-client-py==1.2.6
idna==2.6
itsdangerous==0.24
jieba==0.38
kombu==4.0.2
MarkupSafe==1.0
mutagen==1.38
olefile==0.44
packaging==16.8
Pillow==4.3.0
pkg-resources==0.0.0
pycryptodome==3.4.7
pymongo==3.4.0
PyMySQL==0.7.10
pyparsing==2.2.0
python-alipay-sdk==1.5.1
pytz==2016.10
redis==2.10.5
requests==2.18.4
six==1.10.0
urllib3==1.22
uWSGI==2.0.15
vine==1.1.3
Whoosh==2.7.4
